requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    make upgrade
#
amqp==5.1.1
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
backoff==2.2.1
    # via -r requirements/base.txt
billiard==4.1.0
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.3.4
    # via -r requirements/base.txt
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   pynacl
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==1.5.0
    # via -r requirements/test.in
coverage[toml]==7.3.0
    # via pytest-cov
ddt==1.6.0
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-extensions
    #   django-waffle
    #   edx-django-utils
    #   openedx-filters
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-extensions==3.2.3
    # via -r requirements/base.txt
django-waffle==4.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
edx-django-utils==5.7.0
    # via -r requirements/base.txt
edx-opaque-keys==2.5.0
    # via -r requirements/base.txt
exceptiongroup==1.1.3
    # via pytest
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
kombu==5.3.2
    # via
    #   -r requirements/base.txt
    #   celery
markupsafe==2.1.3
    # via jinja2
newrelic==9.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
openedx-filters==1.6.0
    # via -r requirements/base.txt
packaging==23.1
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.39
    # via
    #   -r requirements/base.txt
    #   click-repl
psutil==5.9.5
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.4.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   celery
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0.1
    # via code-annotations
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.7.1
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   edx-opaque-keys
    #   kombu
tzdata==2023.3
    # via
    #   -r requirements/base.txt
    #   celery
vine==5.0.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
